+++ app/controllers/results_controller.rb
@@ -18,7 +18,7 @@ class ResultsController < ApplicationController
             .includes(:draw, :room, group: %i(lottery_assignment suite))
             .order(:last_name)
     a = %I[#{User.login_attr} last_name first_name draw_name intent group_name
-           lottery_number suite_number room_number]
+           lottery_number building_name suite_number room_number]
     result = CSVGenerator.generate(data: s, attributes: a, name: 'students')
     handle_file_action(**result)
   end